backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langgraph==0.0.20
langchain-openai==0.0.2
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
googlemaps==4.10.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
